+++ packages/pangea-sdk/pangea/services/ai_guard.py
@@ -1,10 +1,154 @@
 from __future__ import annotations
 
-from typing import Any, Dict, Generic, List, Optional, TypeVar, overload
+from typing import Any, Dict, Generic, List, Literal, Optional, TypeVar, overload
 
 from pangea.config import PangeaConfig
 from pangea.response import APIRequestModel, APIResponseModel, PangeaResponse, PangeaResponseResult
 from pangea.services.base import ServiceBase
+
+# This is named "prompt injection" in the API spec even though it is also used
+# for many other detectors.
+PromptInjectionAction = Literal["report", "block"]
+
+MaliciousEntityAction = Literal["report", "defang", "disabled", "block"]
+
+# This is named "PII entity" in the API spec even though it is also used for the
+# secrets detector.
+PiiEntityAction = Literal["disabled", "report", "block", "mask", "partial_masking", "replacement", "hash", "fpe"]
+
+
+class CodeDetectionOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    action: Optional[Literal["report", "block"]] = None
+
+
+class LanguageDetectionOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    allow: Optional[List[str]] = None
+    block: Optional[List[str]] = None
+    report: Optional[List[str]] = None
+
+
+class TopicDetectionOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    block: Optional[List[str]] = None
+
+
+class PromptInjectionOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    action: Optional[PromptInjectionAction] = None
+
+
+class SelfHarmOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    action: Optional[PromptInjectionAction] = None
+    threshold: Optional[float] = None
+
+
+class GibberishOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    action: Optional[PromptInjectionAction] = None
+
+
+class RoleplayOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    action: Optional[PromptInjectionAction] = None
+
+
+class SentimentOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    action: Optional[PromptInjectionAction] = None
+    threshold: Optional[float] = None
+
+
+class MaliciousEntityOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    ip_address: Optional[MaliciousEntityAction] = None
+    url: Optional[MaliciousEntityAction] = None
+    domain: Optional[MaliciousEntityAction] = None
+
+
+class CompetitorsOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    action: Optional[PromptInjectionAction] = None
+
+
+class PiiEntityOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    email_address: Optional[PiiEntityAction] = None
+    nrp: Optional[PiiEntityAction] = None
+    location: Optional[PiiEntityAction] = None
+    person: Optional[PiiEntityAction] = None
+    phone_number: Optional[PiiEntityAction] = None
+    date_time: Optional[PiiEntityAction] = None
+    ip_address: Optional[PiiEntityAction] = None
+    url: Optional[PiiEntityAction] = None
+    money: Optional[PiiEntityAction] = None
+    credit_card: Optional[PiiEntityAction] = None
+    crypto: Optional[PiiEntityAction] = None
+    iban_code: Optional[PiiEntityAction] = None
+    us_bank_number: Optional[PiiEntityAction] = None
+    nif: Optional[PiiEntityAction] = None
+    au_abn: Optional[PiiEntityAction] = None
+    au_acn: Optional[PiiEntityAction] = None
+    au_tfn: Optional[PiiEntityAction] = None
+    medical_license: Optional[PiiEntityAction] = None
+    uk_nhs: Optional[PiiEntityAction] = None
+    au_medicare: Optional[PiiEntityAction] = None
+    us_drivers_license: Optional[PiiEntityAction] = None
+    us_itin: Optional[PiiEntityAction] = None
+    us_passport: Optional[PiiEntityAction] = None
+    us_ssn: Optional[PiiEntityAction] = None
+
+
+class SecretsDetectionOverride(APIRequestModel):
+    disabled: Optional[bool] = None
+    slack_token: Optional[PiiEntityAction] = None
+    rsa_private_key: Optional[PiiEntityAction] = None
+    ssh_dsa_private_key: Optional[PiiEntityAction] = None
+    ssh_ec_private_key: Optional[PiiEntityAction] = None
+    pgp_private_key_block: Optional[PiiEntityAction] = None
+    amazon_aws_access_key_id: Optional[PiiEntityAction] = None
+    amazon_aws_secret_access_key: Optional[PiiEntityAction] = None
+    amazon_mws_auth_token: Optional[PiiEntityAction] = None
+    facebook_access_token: Optional[PiiEntityAction] = None
+    github_access_token: Optional[PiiEntityAction] = None
+    jwt_token: Optional[PiiEntityAction] = None
+    google_api_key: Optional[PiiEntityAction] = None
+    google_cloud_platform_api_key: Optional[PiiEntityAction] = None
+    google_drive_api_key: Optional[PiiEntityAction] = None
+    google_cloud_platform_service_account: Optional[PiiEntityAction] = None
+    google_gmail_api_key: Optional[PiiEntityAction] = None
+    youtube_api_key: Optional[PiiEntityAction] = None
+    mailchimp_api_key: Optional[PiiEntityAction] = None
+    mailgun_api_key: Optional[PiiEntityAction] = None
+    basic_auth: Optional[PiiEntityAction] = None
+    picatic_api_key: Optional[PiiEntityAction] = None
+    slack_webhook: Optional[PiiEntityAction] = None
+    stripe_api_key: Optional[PiiEntityAction] = None
+    stripe_restricted_api_key: Optional[PiiEntityAction] = None
+    square_access_token: Optional[PiiEntityAction] = None
+    square_oauth_secret: Optional[PiiEntityAction] = None
+    twilio_api_key: Optional[PiiEntityAction] = None
+    pangea_token: Optional[PiiEntityAction] = None
+
+
+class Overrides(APIRequestModel):
+    ignore_recipe: Optional[bool] = None
+    """Bypass existing Recipe content and create an on-the-fly Recipe."""
+
+    code_detection: Optional[CodeDetectionOverride] = None
+    competitors: Optional[CompetitorsOverride] = None
+    gibberish: Optional[GibberishOverride] = None
+    language_detection: Optional[LanguageDetectionOverride] = None
+    malicious_entity: Optional[MaliciousEntityOverride] = None
+    pii_entity: Optional[PiiEntityOverride] = None
+    prompt_injection: Optional[PromptInjectionOverride] = None
+    roleplay: Optional[RoleplayOverride] = None
+    secrets_detection: Optional[SecretsDetectionOverride] = None
+    selfharm: Optional[SelfHarmOverride] = None
+    sentiment: Optional[SentimentOverride] = None
+    topic_detection: Optional[TopicDetectionOverride] = None
 
 
 class LogFields(APIRequestModel):
@@ -33,6 +177,8 @@ class AnalyzerResponse(APIResponseModel):
 
 class PromptInjectionResult(APIResponseModel):
     action: str
+    """The action taken by this Detector"""
+
     analyzer_responses: List[AnalyzerResponse]
     """Triggered prompt injection analyzers."""
 
@@ -41,11 +187,13 @@ class PiiEntity(APIResponseModel):
     type: str
     value: str
     action: str
+    """The action taken on this Entity"""
     start_pos: Optional[int] = None
 
 
 class PiiEntityResult(APIResponseModel):
     entities: List[PiiEntity]
+    """Detected redaction rules."""
 
 
 class MaliciousEntity(APIResponseModel):
@@ -58,35 +206,59 @@ class MaliciousEntity(APIResponseModel):
 
 class MaliciousEntityResult(APIResponseModel):
     entities: List[MaliciousEntity]
+    """Detected harmful items."""
+
+
+class CustomEntity(APIResponseModel):
+    type: str
+    value: str
+    action: str
+    """The action taken on this Entity"""
+    start_pos: Optional[int] = None
+    raw: Optional[Dict[str, Any]] = None
+
+
+class CustomEntityResult(APIResponseModel):
+    entities: List[CustomEntity]
+    """Detected redaction rules."""
 
 
 class SecretsEntity(APIResponseModel):
     type: str
     value: str
     action: str
+    """The action taken on this Entity"""
     start_pos: Optional[int] = None
     redacted_value: Optional[str] = None
 
 
 class SecretsEntityResult(APIResponseModel):
     entities: List[SecretsEntity]
+    """Detected redaction rules."""
 
 
 class LanguageDetectionResult(APIResponseModel):
     language: str
     action: str
+    """The action taken by this Detector"""
+
+
+class TopicDetectionResult(APIResponseModel):
+    action: str
+    """The action taken by this Detector"""
 
 
 class CodeDetectionResult(APIResponseModel):
     language: str
     action: str
+    """The action taken by this Detector"""
 
 
 _T = TypeVar("_T")
 
 
 class TextGuardDetector(APIResponseModel, Generic[_T]):
-    detected: bool
+    detected: Optional[bool] = None
     data: Optional[_T] = None
 
 
@@ -94,10 +266,11 @@ class TextGuardDetectors(APIResponseModel):
     prompt_injection: Optional[TextGuardDetector[PromptInjectionResult]] = None
     pii_entity: Optional[TextGuardDetector[PiiEntityResult]] = None
     malicious_entity: Optional[TextGuardDetector[MaliciousEntityResult]] = None
+    custom_entity: Optional[TextGuardDetector[Any]] = None
     secrets_detection: Optional[TextGuardDetector[SecretsEntityResult]] = None
     profanity_and_toxicity: Optional[TextGuardDetector[Any]] = None
-    custom_entity: Optional[TextGuardDetector[Any]] = None
     language_detection: Optional[TextGuardDetector[LanguageDetectionResult]] = None
+    topic_detection: Optional[TextGuardDetector[TopicDetectionResult]] = None
     code_detection: Optional[TextGuardDetector[CodeDetectionResult]] = None
 
 
@@ -112,6 +285,16 @@ class TextGuardResult(PangeaResponseResult, Generic[_T]):
     """Updated structured prompt, if applicable."""
 
     blocked: bool
+    """Whether or not the prompt triggered a block detection."""
+
+    recipe: str
+    """The Recipe that was used."""
+
+    fpe_context: Optional[str] = None
+    """
+    If an FPE redaction method returned results, this will be the context passed
+    to unredact.
+    """
 
 
 class AIGuard(ServiceBase):
@@ -160,7 +343,7 @@ class AIGuard(ServiceBase):
         *,
         recipe: str | None = None,
         debug: bool | None = None,
-        llm_info: str | None = None,
+        overrides: Overrides | None = None,
         log_fields: LogFields | None = None,
     ) -> PangeaResponse[TextGuardResult[None]]:
         """
@@ -180,7 +363,6 @@ class AIGuard(ServiceBase):
                 are to be applied to the text, such as defang malicious URLs.
             debug: Setting this value to true will provide a detailed analysis
                 of the text data
-            llm_info: Short string hint for the LLM Provider information
             log_field: Additional fields to include in activity log
 
         Examples:
@@ -194,7 +376,7 @@ class AIGuard(ServiceBase):
         messages: _T,
         recipe: str | None = None,
         debug: bool | None = None,
-        llm_info: str | None = None,
+        overrides: Overrides | None = None,
         log_fields: LogFields | None = None,
     ) -> PangeaResponse[TextGuardResult[_T]]:
         """
@@ -215,48 +397,10 @@ class AIGuard(ServiceBase):
                 are to be applied to the text, such as defang malicious URLs.
             debug: Setting this value to true will provide a detailed analysis
                 of the text data
-            llm_info: Short string hint for the LLM Provider information
             log_field: Additional fields to include in activity log
 
         Examples:
             response = ai_guard.guard_text(messages=[{"role": "user", "content": "hello world"}])
-        """
-
-    @overload
-    def guard_text(
-        self,
-        *,
-        llm_input: _T,
-        recipe: str | None = None,
-        debug: bool | None = None,
-        llm_info: str | None = None,
-        log_fields: LogFields | None = None,
-    ) -> PangeaResponse[TextGuardResult[_T]]:
-        """
-        Text Guard for scanning LLM inputs and outputs
-
-        Analyze and redact text to avoid manipulation of the model, addition of
-        malicious content, and other undesirable data transfers.
-
-        OperationId: ai_guard_post_v1_text_guard
-
-        Args:
-            llm_input: Structured full llm payload data to be scanned by AI
-                Guard for PII, sensitive data, malicious content, and other data
-                types defined by the configuration. Supports processing up to
-                10KB of JSON text
-            recipe: Recipe key of a configuration of data types and settings
-                defined in the Pangea User Console. It specifies the rules that
-                are to be applied to the text, such as defang malicious URLs.
-            debug: Setting this value to true will provide a detailed analysis
-                of the text data
-            llm_info: Short string hint for the LLM Provider information
-            log_field: Additional fields to include in activity log
-
-        Examples:
-            response = ai_guard.guard_text(
-                llm_input={"model": "gpt-4o", "messages": [{"role": "user", "content": "hello world"}]}
-            )
         """
 
     def guard_text(  # type: ignore[misc]
@@ -264,10 +408,9 @@ class AIGuard(ServiceBase):
         text: str | None = None,
         *,
         messages: _T | None = None,
-        llm_input: _T | None = None,
         recipe: str | None = None,
         debug: bool | None = None,
-        llm_info: str | None = None,
+        overrides: Overrides | None = None,
         log_fields: LogFields | None = None,
     ) -> PangeaResponse[TextGuardResult[None]]:
         """
@@ -286,27 +429,19 @@ class AIGuard(ServiceBase):
                 PII, sensitive data, malicious content, and other data types
                 defined by the configuration. Supports processing up to 10KB of
                 JSON text
-            llm_input: Structured full llm payload data to be scanned by AI
-                Guard for PII, sensitive data, malicious content, and other data
-                types defined by the configuration. Supports processing up to
-                10KB of JSON text
             recipe: Recipe key of a configuration of data types and settings
                 defined in the Pangea User Console. It specifies the rules that
                 are to be applied to the text, such as defang malicious URLs.
             debug: Setting this value to true will provide a detailed analysis
                 of the text data
-            llm_info: Short string hint for the LLM Provider information
             log_field: Additional fields to include in activity log
 
         Examples:
             response = ai_guard.guard_text("text")
         """
 
-        if not any((text, messages, llm_input)):
-            raise ValueError("At least one of `text`, `messages`, or `llm_input` must be given")
-
-        if sum((text is not None, messages is not None, llm_input is not None)) > 1:
-            raise ValueError("Only one of `text`, `messages`, or `llm_input` can be given at once")
+        if text is not None and messages is not None:
+            raise ValueError("Exactly one of `text` or `messages` must be given")
 
         return self.request.post(
             "v1/text/guard",
@@ -314,10 +449,9 @@ class AIGuard(ServiceBase):
             data={
                 "text": text,
                 "messages": messages,
-                "llm_input": llm_input,
                 "recipe": recipe,
                 "debug": debug,
-                "llm_info": llm_info,
+                "overrides": overrides,
                 "log_fields": log_fields,
             },
         )